refactor: impl 형식 사용
