+++ src/main/kotlin/team/incube/flooding/domain/neis/service/impl/GetNeisTimetablesServiceImpl.kt
@@ -1,4 +1,4 @@
-package team.incube.flooding.domain.neis.service
+package team.incube.flooding.domain.neis.service.impl
 
 import com.fasterxml.jackson.databind.JsonNode
 import org.springframework.stereotype.Service
@@ -6,12 +6,13 @@ import team.incube.flooding.domain.neis.client.NeisTimetableClient
 import team.incube.flooding.domain.neis.client.dto.GetTimetablesRequest
 import team.incube.flooding.domain.neis.presentation.data.request.GetNeisTimetablesRequest
 import team.incube.flooding.domain.neis.presentation.data.response.GetNeisTimetablesResponse
+import team.incube.flooding.domain.neis.service.GetNeisTimetablesService
 
 @Service
-class GetNeisTimetablesService(
+class GetNeisTimetablesServiceImpl(
     private val neisTimetableClient: NeisTimetableClient,
-) {
-    fun execute(request: GetNeisTimetablesRequest): GetNeisTimetablesResponse {
+) : GetNeisTimetablesService {
+    override fun execute(request: GetNeisTimetablesRequest): GetNeisTimetablesResponse {
         val response = neisTimetableClient.getTimetables(
             GetTimetablesRequest(
                 officeCode = request.officeCode,
@@ -21,7 +22,6 @@ class GetNeisTimetablesService(
                 date = request.date,
             )
         )
-
         return GetNeisTimetablesResponse(
             date = request.date,
             grade = request.grade,
@@ -31,16 +31,14 @@ class GetNeisTimetablesService(
     }
 
     private fun extractPeriods(response: JsonNode): List<GetNeisTimetablesResponse.Period> {
-        val neisRows =
-            response
-                .path("hisTimetable")
-                .find { node -> node.path("row").isArray }
-                ?.path("row")
+        val neisRows = response
+            .path("hisTimetable")
+            .find { node -> node.path("row").isArray }
+            ?.path("row")
         if (neisRows != null && neisRows.isArray) {
             return neisRows.mapIndexed { index, periodNode ->
                 GetNeisTimetablesResponse.Period(
-                    period =
-                        valueOf(periodNode, "PERIO", "period")?.toIntOrNull() ?: index + 1,
+                    period = valueOf(periodNode, "PERIO", "period")?.toIntOrNull() ?: index + 1,
                     subject = valueOf(periodNode, "ITRT_CNTNT", "subject") ?: "미정",
                     teacher = valueOf(periodNode, "TEACHER_NM", "teacher"),
                     classroom = valueOf(periodNode, "CLRM_NM", "CLASSROOM", "classroom"),
@@ -48,29 +46,25 @@ class GetNeisTimetablesService(
             }
         }
 
-        val targets =
-            listOf(
-                response.path("data"),
-                response.path("timetables"),
-                response,
-            )
+        val targets = listOf(
+            response.path("data"),
+            response.path("timetables"),
+            response,
+        )
 
-        val periodNodes =
-            targets
-                .firstNotNullOfOrNull { node ->
-                    when {
-                        node.isArray -> node
-                        node.path("periods").isArray -> node.path("periods")
-                        node.path("timetables").isArray -> node.path("timetables")
-                        node.path("data").isArray -> node.path("data")
-                        else -> null
-                    }
-                } ?: return emptyList()
+        val periodNodes = targets.firstNotNullOfOrNull { node ->
+            when {
+                node.isArray -> node
+                node.path("periods").isArray -> node.path("periods")
+                node.path("timetables").isArray -> node.path("timetables")
+                node.path("data").isArray -> node.path("data")
+                else -> null
+            }
+        } ?: return emptyList()
 
         return periodNodes.mapIndexed { index, periodNode ->
             GetNeisTimetablesResponse.Period(
-                period =
-                    valueOf(periodNode, "period", "PERIO")?.toIntOrNull() ?: index + 1,
+                period = valueOf(periodNode, "period", "PERIO")?.toIntOrNull() ?: index + 1,
                 subject = valueOf(periodNode, "subject", "ITRT_CNTNT") ?: "미정",
                 teacher = valueOf(periodNode, "teacher", "TEACHER_NM"),
                 classroom = valueOf(periodNode, "classroom", "CLASSROOM", "CLRM_NM"),
@@ -78,15 +72,10 @@ class GetNeisTimetablesService(
         }
     }
 
-    private fun valueOf(
-        node: JsonNode,
-        vararg keys: String,
-    ): String? {
+    private fun valueOf(node: JsonNode, vararg keys: String): String? {
         keys.forEach { key ->
             val value = node.path(key)
-            if (!value.isMissingNode && !value.isNull) {
-                return value.asText()
-            }
+            if (!value.isMissingNode && !value.isNull) return value.asText()
         }
         return null
     }
